re2: build and install static library
e98420fea827155ca9f9f3f20c1830723a0cad6c switched to using cmake, but
cmake only builds the shared (.so) or static library (.a) depending on
whether `BUILD_SHARED_LIBS` is enabled.

To fix this, we run `cmake` twice: once to build the static library,
and once to build the shared library.

Closes #119928.

Signed-off-by: Rui Chen <rui@chenrui.dev>
Signed-off-by: BrewTestBot <1589480+BrewTestBot@users.noreply.github.com>