vnu 22.9.29
There hasn't been a proper release of `vnu` in the GitHub repository
since 20.6.30 (on 2020-06-30) and instead upstream is continually
updating a "latest" release marked as "pre-release". However, they
publish a `vnu-jar` package on npm that includes the jar file we use
in the formula and they *do* version those releases, so we can use
that until upstream creates stable releases on GitHub again.

Closes #120151.

Signed-off-by: Patrick Linnane <patrick@linnane.io>
Signed-off-by: BrewTestBot <1589480+BrewTestBot@users.noreply.github.com>