workflows/tests: do not check other labels if `syntax-only`
Currently, the `setup_tests` job can fail in
`CI-published-bottle-commits` PRs when a `long build` label is present.
That requires us to remove the label and trigger another workflow
manually (or, less ideally, to merge the PRs with failing CI). This may
not be desired because a `long build` label can be useful when we look
up previous PRs.

This was observed in #126454 and #126479.