publish-commit-bottles: check that we can push to the PR
Before trying to checkout the PR branch and pull the bottles, let's
first query the API to see that we will be able to push to the PR
branch.

If we can't, let's bail out early and post a comment asking the user to
allow us to push to their PR branch.

Since we're querying the API already, let's use that data to determine
the remote and branch we want to push to. This will be a lot more
reliable than what `gh pr checkout` generates for us locally. We
currently get the wrong branch name if the PR branch is also named
`master`, for example.

Closes #126975.