tests: determine runners dynamically based on testing formulae
This makes use of `brew determine-test-runners`.

We only deploy a runner for the `tests` job if there are testing formulae
compatible with the given runner. Similarly, we only deploy a runner for
the `test_deps` job if there are dependent formulae compatible with the
given runner to be tested.

This should save time in CI on no-op jobs that take several minutes and
take up space in the queue.