publish-commit-bottles: approve PRs after bottle push
We want to enable the dismissal of reviews after pushes to a PR branch,
but that means that approving reviews are dismissed after @BrewTestBot
pushes the bottle commit(s) to a PR.

To avoid PRs needing a second review before merging, let's have
@BrewTestBot approve the PR after the bottle push.

To support this, I've added a few checks for consistency of the HEAD
SHA1 in the runner's checkout and on the PR branch to avoid merging
commits we don't intend to. These checks are not airtight, but they're
better than nothing. We also check first that the PR is approved before
proceeding.

Finally: I've CCed myself in all comments posted after failures so I can
get a better look at making this workflow handle those situations
better.