fastlane: fix ruby path
The fastlane formula currently depends on the versioned ruby@3.1
formula. However, the PATH environment variable set in the env script
includes the path of the regular ruby formula (currently 3.2). This
leads to confusing behavior where fastlane is using ruby 3.1, but
executing ruby from fastlane will either use the system version or
version 3.2 (depending on whether the regular ruby formula happens to be
installed).

This appears to be a regression in #120124. Previous updates, e.g.
#77459, changed the dependency version and PATH in parallel.