workflows/dispatch-*: set labels separately
Setting them using `gh pr create` seems to trigger a bug where our CI
workflow (`tests.yml`) doesn't see that the CI-published-bottle-commits
label is set. This then results in the bottle fetch step in
`formulae_detect` to not run, and we don't want that.

See [1] for an example of this.

[1] https://github.com/Homebrew/homebrew-core/actions/runs/4734812738/jobs/8404430106