wireguard-tools: append `#{etc}/wireguard` to config paths instead of replace
Replacing `/usr/local/etc/wireguard` with `#{etc}/wireguard` can break other
packages that are relying on `wg-quick` to look in the `/usr/local` path for
.conf files if the user's $HOMEBREW_PREFIX is not `/usr/local`.

This change uses inreplace to append the `#{etc}/wireguard` path to the list of
paths searched rather than replace. It only does this if the `etc` path is
different from `/usr/local`. The end result is that `wg-quick` looks in
either 2 or 3 paths for .conf files:

When $HOMEBREW_PREFIX is `/usr/local`:
- /etc/wireguard
- /usr/local/etc/wireguard

When $HOMEBREW_PREFIX is not `/usr/local` (e.g. /opt/homebrew):
- /etc/wireguard
- /usr/local/etc/wireguard
- /opt/homebrew/etc/wireguard

This maintains compatability with other packages regardless of the user's
$HOMEBREW_PREFIX location.