workflows: prevent premature removal of long timeout label
Currently, it is possible to remove the long timeout label from a PR
before CI on that PR finishes. This is circumvents the limit we have on
the number of long timeout PRs running at a time.

This workflow prevents that from happening. It works by re-adding the
long timeout label upon removal, and then triggers our long timeout
label removal workflow which checks the state of CI to determine whether
removal is appropriate, and removes it if it is.

This will not cause an infinite loop of workflows triggering each other,
since we use `@github-actions` to remove the long timeout label, and
`@github-actions` does not trigger workflow events.