publish-commit-bottles: remove check for reviews requesting changes
When a reviewer requests changes but later approves a PR, their original
review requesting changes still shows a state of `CHANGES_REQUESTED`, so
we mistakenly error out here currently.

Let's loosen this to check only for any approving review. We could also
make the check smarter (e.g. check that a reviewer requesting changes
later gives an approving review), but that is a little tricky to do with
only Bash and jq.