jsoncpp: add a note on the deprecation of the CMake build
We've had multiple PRs to try to switch this to CMake (#103386, #133342)
but we don't want to do that because the CMake build is deprecated.
Let's leave a comment here about that to avoid accidentally making this
change in the future.

While we're here, let's update the `install` method to use the newer
style.

Closes #133342.