mosh: run `make check`
The `protobuf` formula forces us to build this with a newer C++ standard
than upstream expects or even tests, so let's run `make check` to ensure
that we didn't break anything.

Running the tests takes only about a minute. Also, since this software
is essentially a remote shell, it is cryptography-adjacent, which is
close to our requirement for formulae to run `make check`.