google-authenticator-libpam: fix qr code generation
Before this change, google-authenticator prints out this warning about
not being able to use `libqrencode`:

    $ google-authenticator --time-based
    Warning: pasting the following URL into your browser exposes the OTP secret to Google:
      https://www.google.com/chart?chs=200x200&chld=M|0&cht=qr&chl=otpauth://totp/jeremy@Jeremys-MBP.lan%3Fsecret%3DSLP5Q45AMDZJILWEF6PA6V7PO4%26issuer%3DJeremys-MBP.lan
    Failed to use libqrencode to show QR code visually for scanning.
    Consider typing the OTP secret into your app manually.
    Your new secret key is: SLP5Q45AMDZJILWEF6PA6V7PO4
    Enter code from app (-1 to skip):

Note how there was no rpath in the `google-authenticator` binary:

    $ otool -l $(which google-authenticator) | grep -A 2 RPATH
    $

After this change, I get a pretty qr code when I run
`google-authenticator`. Note the RPATH in the binary:

    $ otool -l $(which google-authenticator) | grep -A 2 RPATH
              cmd LC_RPATH
          cmdsize 48
             path /opt/homebrew/opt/qrencode/lib (offset 12)

I have no idea if this is the right way to fix this issue. I copied this
approach with an `LDFLAGS` env var from
[`Formula/libgit2-glib.rb`](https://github.com/Homebrew/homebrew-core/blob/0d0aea5374bb2250b76cacdd0e4de69237d25508/Formula/libgit2-glib.rb).