libgccjit: fix CI installation failure
The gcc build evades our superenv shims, so it is not linked with
`-headerpad_max_install_names`. This results in `brew install` failures
in CI when we try to pour a bottle with relocatable install names.

See, for example, https://github.com/Homebrew/homebrew-core/actions/runs/5724512846/job/15517640743#step:3:343

We can fix that by passing the right arguments to `make`.