create-replacement-pr: approve and merge applicable PRs
Let's automatically approve replacement PRs where the original PR was
approved, and enable auto-merge on replacement PRs where we've already
uploaded bottles for.

As in our other workflows, we need to use `GITHUB_TOKEN` to approve
(because the replacement PR is created by @BrewTestBot) and
`HOMEBREW_GITHUB_PUBLIC_REPO_TOKEN` to enable auto-merge (so that our
merge queue CI runs).