ffmpeg: patch for QtWebEngine / chromium
Fixes build of chromium in QtWebEngine

Ideally the chromium code should be modified to not rely on
a private method, and I will see if I can open another issue
upstream to get rid of the ffmpeg patch in the long term.

We should merge this nevertheless as this will allow to build QtWebEngine on Linux,
and rebuild a PyQt bottle.

The patch has already been accepted in arch Linux