quilt: add patch to fix stable on macOS
This adds a patch which is already upstream in the tool's repository but
not included in a stable version. Without this fix, refresh subcommand
of quilt is broken on macOS/BSD-based system due to awk. It leads to
incorrect behavior which just appends the whole patch file content on
the file again and again on each call.

This patch should be removed on next release.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>