cling: deprecate
Does not build on Ventura
I asked for a new release in May 2023, the response was not super clear:
https://github.com/root-project/cling/issues/492#issuecomment-1555938334

The comments on https://github.com/Homebrew/homebrew-core/pull/131473
did also not really look helpful.

Download count is low, so deprecating is the best thing to do, and
gives upstream some more time to react before removal on our side.