cargo-bloat: improve test
Currently, in the `test` block, both stable and beta toolchains are
installed, while only the beta is used. (The stable toolchain is
installed by default during the `rustup-init` invocation.) We can
instead ask `rustup-init` to install the beta toolchain directly.

In addition, we can choose to use the minimal profile, so that only
the necessary components (`rustc`, `rust-std`, and `cargo`) are
installed [^1]. (The default profile additionally installs `rust-docs`,
`rustfmt`, and `clippy`, but we don't use them in the test.)

That should help to cut the test time by half.

[^1]: https://rust-lang.github.io/rustup/concepts/profiles.html