apprise 1.5.0 (new formula)
Adding a new formula for apprise. This is a python tool generated with the brew create --python command and worked without changes. Tests are added to make sure notifications work by creating a custom notifier that just prints to the command line. The output of the command is then checked to look for the expected value.

apprise 1.5.0 (new formula)

Adding a new formula for apprise. This is a python tool generated with the brew create --python command and worked without changes. Tests are added to make sure notifications work by creating a custom notifier that just prints to the command line. The output of the command is then checked to look for the expected value.