libplacebo: use expected resource directory names
The `libplacebo` build is currently failing due to certain resources
not being found in the expected directories (e.g., `glad2` should be
in `3rdparty/glad`, not `3rdparty/glad2`). Modifying the resource
names will lead to an audit failure, so we have to massage the names
in the `install` method to fix the build.