audiofile: update resource url to fix test
The test for `audiofile` uses the formula's `aiff` resource but the
www-mmsp.ece.mcgill.ca URL leads to a certificate error. This issue
occurs because the certificate is for mmsp.ece.mcgill.ca instead.

This addresses the issue by updating the resource URL to use
mmsp.ece.mcgill.ca, allowing the test to work again.