libxmi: fix build on newer macOS versions
Building `libxmi` on newer macOS versions can lead to errors with
notes like `include the header <stdlib.h> or explicitly provide a
declaration for 'exit'`. In the `configure` file, `<stdlib.h>` and
`<stddef.h>` are included when `STDC_HEADERS` is defined and this
happens when `$ac_cv_header_stdc = yes`.

That is not the case on newer macOS versions, so we need to either
manually set this in `ENV` or simply run `autoreconf` like we do on
Linux.