libvpx: add patch to handle Sonoma targets
As explained by Bo, the build failures that we were seeing arose
because `libvpx` hasn't added support for Sonoma targets yet
(`arm64-darwin23-gcc`, `x86_64-darwin23-gcc`), so it was being built
as a Linux library instead. If we patch the `configure` and
`build/make/configure.sh` files to support the new targets, the build
works as expected without the previous workarounds.

This is something that we will need to do each year if upstream
hasn't patched these files before we try to bottle `libvpx` for a new
macOS version. I've added a comment to explain the situation and the
addition of the `--target` arg will produce a more meaningful error
(e.g., `Unrecognized toolchain 'arm64-darwin23-gcc'`), so hopefully
this is clearer next time.

Co-authored-by: Bo Anderson <mail@boanderson.me>