rpm 4.19.0
Also, make this Linux-only. It requires Linux-only `elfutils`, relies
heavily on GNU extensions like `strchrnul` from `strings.h` and
`GLOB_ONLYDIR` from `glob.h`, and has less (if any) practical usage
outside of Linux. Many of the above requirements used to be conditional
when Autotools was used. But now with CMake this is just not possible to
build without extensive patching.

Co-authored-by: Sean Molenaar <SMillerDev@users.noreply.github.com>
Co-authored-by: Ruoyu Zhong <zhongruoyu@outlook.com>