wbox: disable
hping.org has been unavailable since at least 2023-02-15 (the server
does not respond) and the `wbox` formula has been deprecated as a
result. This disables the formula, as it has been more than six months
since it was disabled.

Besides that, this updates the `homepage` and `stable` URLs to use
the last available archive.org snapshots and adds license information.