whalebrew: update build version
Context
---

As raised in whalebrew/whalebrew#233, the current version number
embedded in the whalebrew binary is incorrect.

Goal
---

This commit aims at ensuring that the whalebrew version displayed
running `whalebrew version` represents the actual version of the
whalebrew code.

As per the [semver] semantics, a semver version may include build
informations, following this patterns:

```
<version core> "+" <build>
<build> ::= <dot-separated build identifiers>
<dot-separated build identifiers> ::= <build identifier>
                                    | <build identifier> "." <dot-separated build
<build identifier> ::= <alphanumeric identifier>
                     | <digits>
```

Benefit this usage to mention reflect that the whalebrew binary was
compiled by homebrew and not taken from the original whalebrew release.

[semver]: https://semver.org/

Fixes: whalebrew/whalebrew#233