jam: update stable, fix undeclared function errors
The existing `stable` URL for `jam` returns a 404 (Not Found) error,
as the file has moved to a different location. This updates the URL
to the current file location.

This also adds a patch to fix a couple of `undeclared function` errors
that have surfaced on macOS Sonoma.

Lastly, this adds a `livecheck` block to skip the formula, as it is
no longer developed.

Co-authored-by: Rui Chen <rui@chenrui.dev>