antiword: update homepage, stable
The `homepage` and `stable` URLs in the `antiword` formula have been
returning 404 (Not Found) responses for some time but we can't
deprecate the formula because `textract` depends on it. We can't
bottle this for Sonoma due to the broken www.winfield.demon.nl URLs,
so this updates the URLs to their last good archive.org snapshots.

Co-authored-by: Patrick Linnane <patrick@linnane.io>