deno: remove `llvm` from library search paths
There is an issue with the Xcode 15 linker which causes linkage against
LLVM's libunwind due to it being present in a library search path. The
extra linkage was caught by our linkage check, and as a result, we could
not bottle it (as well as several other formulae having LLVM as
build-time dependency) for Sonoma. This serves as a temporary workaround
for the issue.

See also: llvm/llvm-project@7fcbb64fca5e664b11f9167a215b47de2ae9082c