libdicom 1.0.5 (new formula)
OpenSlide 4.0 will be released within the next 10 days, and has support
for DICOM slides via libdicom as a major new feature and a required
dependency.

libdicom is a new library developed with sponsorship from NIH. It is a
small, fast, free C library with no dependencies, designed to make reading
DICOM WSI images as fast as the vendor formats it replaces.

https://github.com/ImagingDataCommons/libdicom

Even though this repo does not yet have the number of stars homebrew
asks for, we're hoping we can get it accepted, since openslide4 won't
build without it.

libdicom: update test

Signed-off-by: Rui Chen <rui@chenrui.dev>