argyll-cms: vendor jam
We would like to get rid of jam, as it should be deprecated.
The only formula that uses jam is argyll-cms
Upstream confirmed that they wanted to continue to use jam.

Thus, the fate of both tools are now tied. This change will allow to get
rid of the jam formula.

As long as argyll-cms builds without too much patching all should be good ...