xmlto: drop nonet patch
libxml2 was misconfigured so the xml catalog didn't work.
To work around that problem, xmlto was patched to allow downloading
xmls from internet instead. That fails if the internet access is disabled.
But libxml2 was fixed by commit 02ebdcde61.
So drop the no longer needed patch, as with libxml2 fix, catalog
works locally.