workflows/triage: don't check commit format for commits pushed by @BrewTestBot
The commit style check is run on every push, but skipped when the
`CI-published-bottle-commits` label is added. However, since we add the
label after pushing the bottle commits, it is possible that commit
style check is still run on the PR (which would then show a failing
status because it determines that the commits are not squashed).

Let's work around that by skipping the check when the commits are pushed
by @BrewTestBot. We trust that @BrewTestBot will always remember to use
the correct commit format.