workflows: Support CI-version-downgrade label
`brew audit` and test-bot now support skipping the "stable version
should not decrease" audit, so this modifies homebrew/core CI to
control this behavior using a `CI-version-downgrade` label. This will
allow us to selectively skip this check when we need CI to run on a
version downgrade PR (e.g., to produce new bottles).