tarsnap: backport fix for "scrypt code is broken"
The current tarsnap 1.0.40 (lastest stable release) binary does not
allow users to have passphrase-protected keyfiles on some CPU
architectures, due to a bug in strict aliasing.  Any attempt to do so
results in the message:

    tarsnap: Generic scrypt code is broken - please report bug!

This was fixed in the upstream repository on 2022-04-01 [*], but there
has not been a stable release since then.

When tarsnap 1.0.41 is released, this commit can be reverted.

[*] https://github.com/Tarsnap/tarsnap/commit/4af6d8350ab53d0f1f3104ce3d9072c2d5f9ef7a

Closes #160754