Add build-provenance after pr-pull
Adds an extra step after pr-pull in publish-commit-bottles.yml to
generate build provenance for bottles that get published.

This requires a small change in `Homebrew/brew` to retain the temporary
directory and to expose the path to the following workflow steps. That
change must be merged first before this change will work.